created utils file and changes made
